requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.2
beanie==1.23.6
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.1.0
